pyproject.toml
[project]
name = "yolo-brand-detector"
version = "0.1.0"
description = "Two-stage YOLO v11 trainer for brand detection"
authors = [
    {name = "AI Squad", email = "team@aisquad.com"}
]
dependencies = [
    "ultralytics>=8.2.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "Pillow>=10.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "pandas>=2.0.0",
    "tqdm>=4.65.0",
    "PyYAML>=6.0",
    "tensorboard>=2.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
